Created a seperate category crud controller
